# Conflicts:
#	src/main/java/com/task/comments/services/CommentService.java
#	src/main/java/com/task/posts/dto/PostMapper.java

+++ src/main/java/com/task/posts/service/PostService.java
@@ -89,6 +89,8 @@ public class PostService {
     }
 
     public void deletePost(Long id){
+        if(postDao.getPostById(id).isEmpty())
+            throw new ResourceNotFound("Post not found");
         postDao.deletePost(id);
     }
